+++ composer.json
@@ -11,7 +11,7 @@
     ],
     "require": {
         "php": "^8.1",
-        "vatly/vatly-api-php": "^0.1.0-alpha.18"
+        "vatly/vatly-api-php": "^0.1.0-alpha.19"
     },
     "require-dev": {
         "mockery/mockery": "^1.6",
