+++ docs/webhook-flow.md
@@ -1,8 +1,8 @@
 # Webhook flow: event → reaction → repo method → fields
 
-A driver author shouldn't have to read four directories (the event DTOs in `vatly-api-php`'s `Vatly\API\Webhooks\Events\*`, `src/Webhooks/Reactions/`, `WebhookEventFactory`, `src/Contracts/`) to answer *"if I implement `OrderWriter::store`, which webhooks flow into it and with what data?"* This page is that single reference.
+A driver author shouldn't have to read several directories (the event DTOs and the `WebhookEventFactory` in `vatly-api-php`'s `Vatly\API\Webhooks\*`, plus fluent's `src/Webhooks/Reactions/` and `src/Contracts/`) to answer *"if I implement `OrderWriter::store`, which webhooks flow into it and with what data?"* This page is that single reference.
 
-> **Where the event DTOs live:** the typed webhook events (`OrderPaid`, `OrderPaymentFailed`, `SubscriptionStarted`, …) are owned by **`vatly-api-php`** under `Vatly\API\Webhooks\Events\*` and consumed by fluent — fluent no longer ships its own `src/Events/` copies (only the driver-side `SubscriptionWasCreatedFromWebhook` / `OrderWasCreatedFromWebhook` / `NullEventDispatcher` remain in `Vatly\Fluent\Events\`). Their `taxSummary` field is a `Vatly\API\Types\TaxSummaryCollection` whose items expose `taxRate` (a `TaxSummaryRate`) + `amount` (a `Money`); call `$item->amount->toCents()` for integer cents. Order lines are `Vatly\API\Types\OrderLineData[]` (note: `Vatly\API\Types\`, not the old `Vatly\API\Data\`), and each line's `basePrice` / `total` / `subtotal` are likewise `Money`. The orchestration (`WebhookEventFactory`, the reactions, `WebhookProcessor`) stays in fluent.
+> **Where the event DTOs live:** the typed webhook events (`OrderPaid`, `OrderPaymentFailed`, `SubscriptionStarted`, …) are owned by **`vatly-api-php`** under `Vatly\API\Webhooks\Events\*` and consumed by fluent — fluent no longer ships its own `src/Events/` copies (only the driver-side `SubscriptionWasCreatedFromWebhook` / `OrderWasCreatedFromWebhook` / `NullEventDispatcher` remain in `Vatly\Fluent\Events\`). Their `taxSummary` field is a `Vatly\API\Types\TaxSummaryCollection` whose items expose `taxRate` (a `TaxSummaryRate`) + `amount` (a `Money`); call `$item->amount->toCents()` for integer cents. Order lines are `Vatly\API\Types\OrderLineData[]` (note: `Vatly\API\Types\`, not the old `Vatly\API\Data\`), and each line's `basePrice` / `total` / `subtotal` are likewise `Money`. The event construction — verify → parse → map — is owned by **`vatly-api-php`**: `Vatly\API\Webhooks\WebhookEventFactory` builds the typed events from the signed payload. Fluent owns the downstream orchestration (the reactions and `WebhookProcessor`, which records → reacts → dispatches) and consumes that factory.
 
 > **Event money is `Money`, not int cents (api-php ≥ `0.1.0-alpha.18`):** the `total` / `subtotal` fields on the money-bearing events are `Vatly\API\Types\Money` value objects (decimal-string `value` + `currency`), not raw integer cents. Read the currency off the Money (`$event->total->currency`) and flatten to integer cents with `$event->total->toCents()`. The **standalone `currency` field was removed** from `OrderPaid`, `OrderPaymentFailed`, `RefundCompleted`, `RefundFailed`, and `RefundCanceled` (their `total` / `subtotal` are non-null `Money`). The chargeback events `OrderChargebackReceived` / `OrderChargebackReversed` are the exception: their `total` / `subtotal` are typed **nullable** (`?Money`) at the api-php contract level and they **keep** a standalone `currency` string — though the fat chargeback payload populates the amounts in practice. Fluent's built-in reactions flatten `Money → int` at the persistence edge, so the driver-facing `Store*Data` / `Update*Data` DTOs still carry integer-cents `total` / `subtotal` and a `currency` string — drivers are unaffected.
 
@@ -10,7 +10,7 @@ The [README's "Webhook pipeline at a glance"](../README.md#webhook-pipeline-at-a
 
 ## How events are built (no API round-trip)
 
-Vatly sends **fat, HMAC-signed** webhook deliveries: the payload's `object` is the full resource — byte-identical to the corresponding `GET /…/{id}` body (subtotal, the complete tax summary, lines, mandate). The HMAC signature (verified in `WebhookProcessor`) is the trust boundary, so the payload is the authoritative snapshot and there is **no follow-up API GET to "enrich"**:
+Vatly sends **fat, HMAC-signed** webhook deliveries: the payload's `object` is the full resource — byte-identical to the corresponding `GET /…/{id}` body (subtotal, the complete tax summary, lines, mandate). The HMAC signature (verified in `WebhookProcessor`) is the trust boundary, so the payload is the authoritative snapshot and there is **no follow-up API GET to "enrich"**. The mapping below is performed by api-php's `Vatly\API\Webhooks\WebhookEventFactory`:
 
 - **`order.paid`, `order.payment_failed`, `refund.*`, `order.chargeback_*`, `subscription.started`, `subscription.billing_updated`** — the matching api-php Resource (`Order` / `Refund` / `Chargeback` / `Subscription`) is hydrated in memory from `$webhook->object` (via `ResourceFactory::createResourceFromApiResult()`, the same path the API client uses to build a resource from a `GET` response), then mapped through the event's `fromApi*` constructor. Same enriched event — full tax breakdown, lines, mandate, dispute status — with zero network call.
 - **`order.canceled`, `checkout.*`, `subscription.cancellation_grace_period_completed`, `webhook.setup`** — built straight from the payload envelope; they carry no money/tax resource to hydrate.
